refactor: use refresh token own file
