+++ packages/shared/src/contexts/BootProvider.tsx
@@ -3,12 +3,10 @@ import React, {
   ReactNode,
   useCallback,
   useEffect,
-  useRef,
   useState,
 } from 'react';
 import { useQuery, useQueryClient } from 'react-query';
-import { differenceInMilliseconds } from 'date-fns';
-import { AccessToken, BootCacheData, getBootData } from '../lib/boot';
+import { BootCacheData, getBootData } from '../lib/boot';
 import FeaturesContext from './FeaturesContext';
 import { AuthContextProvider } from './AuthContext';
 import { AnonymousUser, LoggedUser } from '../lib/user';
@@ -20,30 +18,7 @@ import {
   themeModes,
 } from './SettingsContext';
 import { storageWrapper as storage } from '../lib/storageWrapper';
-
-export function useRefreshToken(
-  accessToken: AccessToken,
-  refresh: () => Promise<unknown>,
-): void {
-  const timeout = useRef<number>();
-
-  useEffect(() => {
-    if (accessToken) {
-      if (timeout.current) {
-        window.clearTimeout(timeout.current);
-      }
-      const expiresInMillis = differenceInMilliseconds(
-        new Date(accessToken.expiresIn),
-        new Date(),
-      );
-      // Refresh token before it expires
-      timeout.current = window.setTimeout(
-        refresh,
-        expiresInMillis - 1000 * 60 * 2,
-      );
-    }
-  }, [accessToken, refresh]);
-}
+import { useRefreshToken } from '../hooks/useRefreshToken';
 
 export const BOOT_LOCAL_KEY = 'boot:local';
 export const BOOT_QUERY_KEY = 'boot';
